Installer: Exclude all .pdb files, except Halyard.pdb by default
This patch adds *.pdb to the list of files to exclude from engine/win32 in
our windows-installer-template.iss.  It also adds a flag to optionally
include Halyard.pdb, which is highly useful for debugging crashes in the
field; that flag defaults to true, but can be overridden in each program's
config/windows-installer.iss.