Exclude test media directories from installer and updater
In IML libs, we have a couple of directories of test media that are fairly
large.  Since our SAE update is fairly slow right now, and Joe would very
much like it to be faster, we need to try and eliminate as much extra
download overhead as we can.  This patch adds excludes for those test
media directories.  It also updates buildscript to be able to support
such excludes in our Inno Setup parser that is used for building our
update manifests.