UpdateInstaller: Extract InstallTool superclass, add new Uninstaller class
In order to support uninstallation and cleanup of the install directory
in preparation for a new installation, we will need to be able to perform
installation related tasks on just a destination directory, without an
update source directory.  This patch factors out a superclass, InstallTool,
from UpdateInstaller, as well as creating a new subclass Uninstaller which
will be responsible for uninstallation and cleanup before a new install.