UpdateInstaller: Pull more functionality up into InstallTool
This patch pulls the functionality for preparing for an update,
checking if it is possible, and running the update up into the InstallTool
class, and bases the existing uninstaller code on it.  We have not yet
factored out the cleanup code that will actually be necessary to run
the uninstallation cleanup.

Note that in order to get the various pieces to work, we've had to add
a hack to write the log file out to the directory we're uninstalling from
during an uninstall.  This is less than ideal, but we'll have to accept
it for now until we can find a better location for it or disable the
logger when doing an uninstall.