UpdateInstaller: Perform cleanup from Uninstall
This patch adds the code to actually do a full cleanup pass during uninstall,
in order to delete all of the files that the updater knows about that
InnoSetup does not, and thus cannot uninstall.

In order to do this, we call BuildCleanupRecursive from the
Uninstaller class.  We tell it to clean out everything in our
manifests, and all directories, but ignore the manifests themselves,
release.spec, the temp directory, and the root installation directory
itself.

In order to be able to construct arguments for BuildCleanupRecursive,
I would have needed to construct a few FileSet::LowercaseFilenameMaps.
In order to do that, though, I would have needed manifest entries to
put in those maps.  Since none of the code involved actually depended
on the entries, I refactored the code to combine
UpdateInstaller::FilenameSet, UpdateInstaller::DirectoryNameMap, and
FileSet::LowercaseFilenameMap into
UpdateInstaller::LowercaseFilenameMap.  It now maps simply from the
lowercase name to the case-preserved name, like DirectoryNameMap did,
and FileSet does not need to be involved any more.  FilenameSet had
only been used as an intermediate type for computing the
DirectoryNameMaps, so it has been eliminated.