UpdateInstaller: Look in correct directory for Halyard to launch when done
Back when we fixed the updater to work out of two separate
directories, a download directory and the installed program directory,
we didn't bother updating the unit tests to split the fixture
directory in two, since it worked just fine to pass the same directory
in for each argument.  Now that we are attempting to clean up the
install directory before uninstall, it causes problems to have that extra
Update directory in our installed directory.  This patch thus splits the
fixture directory for the old unit tests into a download-dir and an
installed-program.

In doing so, I discovered a bug; our code that is supposed to launch
Halyard if it can't find the program passed in (which will be Tamale
during the double update) was looking for Halyard in the download
directory, instead of the installed directory.  This patch fixes that
problem, as it is necessary to get the unit tests to pass once we have
done the reorg mentioned above.