Build fix: Disable precompiled headers on new UpdateInstaller files
When I created the InstallTool.cpp and Uninstaller.cpp files, I turned
off precompiled header compilation in the Debug build, but not the release
build.  This meant that 'rake test' would run in the UpdateInstaller
directory, as it only builds the Debug build, but 'rake test' failed in
the top-level Halyard directory, as it builds both Debug and Release builds.
This patch disables precompiled headers for these files in the Release build
as well.

Fix-for: a41a73c779930ab149c403220d22ae2a7c082434