UpdateInstaller: Add better logging of failed updates and uninstalls
In order to help debug failed updates and uninstalls, we log all phases of
the uninstall process (where formerly we omitted logging of many phases
during uninstall).  We also add reasons why the update or uninstall is
impossible, if we determine that it is impossible during our IsPossible
check; to do this, we add a Describe method to each FileOperation class,
and log the description of the impossible file operation that prevents
us from running if there is one.  This now requires that LogFile be
initialized in the test case that tests an impossible update, so we
add that initialization to that test case.