UpdateInstaller: Don't try to delete UpdateInstaller.exe during uninstall
We cannot delete UpdateInstaller.exe while uninstalling, since it's currently
running.  So, we add that to our list of files to keep.  We also need to
fix BuildCleanupRecursive to allow overriding explicit deletes (which in
uninstall, are computed based on the installed manifest), by providing
explicit keeps.

This patch also adds a few more files to the list of files to explicitly
keep (the lock file, the install log) as an extra precaution and to reduce
the number of warnings about undeletable files we get in our logs.  It also
constructs the list of directories to keep from the list of files to keep,
instead of building it manually, so that those will be able to stay in
sync better if we need to add more files and directories later.