UpdateInstaller: Create lock file if possible during uninstall
In order to reduce the chances of winding up with a half-working program,
we put a lock file in our directory during update installs to prevent
other updates from happening concurrently, and to prevent the program
from running if the update stops half-way through.  This patch adds the
same lock file during the uninstall process; however, the presence of the
lock file to begin with doesn't prevent uninstall (as a leftover lock file
from a failed update shouldn't prevent uninstall).