UpdateInstaller: Remove AUTO-UPDATE from manifests
Our AUTO-UPDATE file is supposed to be an empty file that the installer
puts in place to indicate that programs should automatically update;
if absent, automatic update will not happen and users will need to
manually update.  Because its presence is supposed to be controlled
by the installer, it shouldn't ever be touched by the updater; we don't
want to enable automatic updates as soon as someone has done a single
manual update.  We made a mistake a while back, however, and included
it in the 'base' component, which caused it to be included in the
corresponding manifest.

So, this patch removes it from that component.  Our updater should be
intelligent enough to simply ignore a file that it thinks needs to be
deleted but actually isn't present on disk, so the case of someone who
has automatic updates disabled should be covered by just removing it
from the component.  If people have auto-update turned on, however,
removing this file now would delete it from their disk, so we need to
add code to the update installer to specifically ignore the
AUTO-UPDATE file in order to avoid disabling automatic updates for
them.