UpdateInstaller: If file needed from tree does not exist, keep looking
Now that we have the smarter updater that does not download files that
it believes exist on disk, we have run into a problem in certain buggy
cases like that of AUTO-UPDATE.  AUTO-UPDATE is a file that is present
in the manifest, but which may actually be absent from the disk.  It
happens to be the empty file, which is needed in several other places
in the tree. Luckily, there are other copies on disk as well, but the
update installer was always just picking the first.  This patch adds
an existence check when picking the file to copy or move from the
existing tree; if the necessary file doesn't exist, it keeps looking
to see if it exists elsewhere.