Installer: Uninstall an existing program before installing
Because we have done a significan reorganization of the Halyard
directory layout between the 0.4.x release series and the 0.6.x
release series, we need to clean out our old files and directories
so they won't cause problems to the new program.  Inno Setup does
not delete old files, nor does it rename directories that differ
only in case, which would leave us in a broken state.

In order to do this, we simply perform an uninstall, deleting
all files that we have listed in our manifests, and clearing
out all directories that are empty after doing so.  If there are
files or directories that for some reason prevent us from doing
an uninstall, we fail and tell the user that they need to
choose a different install directory or clean out the directory
manually.

Due to a problem in which our program would exit with a result
code of 0 in certain cases when calling abort(), we have switched
our fatal error handler to simply call exit(1) instead, so the
installer can properly check its result code.