UpdateInstaller: Distinguish --cleanup and --uninstall cases
We are using the same code to clean out a directory during
uninstallation, adn to clean out a directory before installing a new
program.  These two use cases actually have different requirements,
though.  When we are uninstalling, we just want to get rid of all of
the files we know that we control, to leave the user with as clean a
system as we can give them.  When doing cleanup before installing a
new program, though, we want to make sure we have cleaned out
everything that might cause problems, so stray files and directories
don't cause problems for the new install.  If that is not possible,
however, we want to leave their existing directory untouched,
detecting the problem before we try to proceed.

The behavior described for the --cleanup option is how the --uninstall
option has been implemented until now.  This patch distinguishes these
two options, and simply skips the IsPossible check in the --uninstall
case, as well as catching errors while performing our uninstall
operations and simply logging them rather than letting them crash the
program.