buildscript: inno_setup.rb: Allow version checking preprocessor directives
We need the ability to add Inno Setup version checks to our installer
template files.  This patch adds support for Inno Setup preprocessor
version checks, though for our implementation, it always considers the
version to be greater than the one checked against.  It also adds an
implementation of the #error directive, so that we can ignore it if we
encounters it in an inactive state.